refactor(urlcanon): make shared URL canonicalizer its own go 1.25.0 module
Extract the AdCP URL-identifier canonicalizer into a top-level
github.com/adcontextprotocol/adcp-go/urlcanon module declaring go 1.25.0.
The adcp module (go 1.26.2) and the root module both consume it via a local
replace, so no app module's Go version is bumped: every pre-existing go.mod
keeps its origin/main go directive (adcp 1.26.2, everything else 1.25.0).
