+++ adcp/go.mod
@@ -4,6 +4,7 @@ go 1.26.2
 
 require (
 	github.com/DATA-DOG/go-sqlmock v1.5.2
+	github.com/adcontextprotocol/adcp-go/urlcanon v0.0.0
 	github.com/google/jsonschema-go v0.4.2
 	github.com/gowebpki/jcs v1.0.1
 	github.com/modelcontextprotocol/go-sdk v1.5.0
@@ -20,3 +21,5 @@ require (
 	golang.org/x/sys v0.41.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
+
+replace github.com/adcontextprotocol/adcp-go/urlcanon => ../urlcanon
