+++ cmd/context-agent/go.mod
@@ -1,6 +1,6 @@
 module github.com/adcontextprotocol/adcp-go/cmd/context-agent
 
-go 1.26.2
+go 1.25.0
 
 require (
 	github.com/adcontextprotocol/adcp-go v0.0.0
@@ -10,7 +10,7 @@ require (
 )
 
 require (
-	github.com/adcontextprotocol/adcp-go/adcp v0.0.0 // indirect
+	github.com/adcontextprotocol/adcp-go/urlcanon v0.0.0 // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/go-logr/logr v1.4.3 // indirect
@@ -46,4 +46,4 @@ replace github.com/adcontextprotocol/adcp-go/registry => ../../registry
 
 replace github.com/adcontextprotocol/adcp-go/registry/redisstore => ../../registry/redisstore
 
-replace github.com/adcontextprotocol/adcp-go/adcp => ../../adcp
+replace github.com/adcontextprotocol/adcp-go/urlcanon => ../../urlcanon
