+++ e2e/go.mod
@@ -1,6 +1,6 @@
 module github.com/adcontextprotocol/adcp-go/e2e
 
-go 1.26.2
+go 1.25.0
 
 require github.com/adcontextprotocol/adcp-go v0.0.0
 
@@ -19,5 +19,3 @@ require (
 )
 
 replace github.com/adcontextprotocol/adcp-go => ../
-
-replace github.com/adcontextprotocol/adcp-go/adcp => ../adcp
