+++ reference/context-agent/go.mod
@@ -1,6 +1,6 @@
 module github.com/adcontextprotocol/adcp-go/reference/context-agent
 
-go 1.26.2
+go 1.25.0
 
 require (
 	github.com/adcontextprotocol/adcp-go v0.0.0
@@ -9,7 +9,7 @@ require (
 )
 
 require (
-	github.com/adcontextprotocol/adcp-go/adcp v0.0.0 // indirect
+	github.com/adcontextprotocol/adcp-go/urlcanon v0.0.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
@@ -23,6 +23,5 @@ require (
 replace (
 	github.com/adcontextprotocol/adcp-go => ../../
 	github.com/adcontextprotocol/adcp-go/targeting/prommetrics => ../../targeting/prommetrics
+	github.com/adcontextprotocol/adcp-go/urlcanon => ../../urlcanon
 )
-
-replace github.com/adcontextprotocol/adcp-go/adcp => ../../adcp
