+++ targeting/prommetrics/go.mod
@@ -1,11 +1,11 @@
 module github.com/adcontextprotocol/adcp-go/targeting/prommetrics
 
-go 1.26.2
+go 1.25.0
 
 require github.com/adcontextprotocol/adcp-go v0.0.0
 
 require (
-	github.com/adcontextprotocol/adcp-go/adcp v0.0.0 // indirect
+	github.com/adcontextprotocol/adcp-go/urlcanon v0.0.0 // indirect
 	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
 	golang.org/x/crypto v0.51.0 // indirect
@@ -22,4 +22,4 @@ require (
 
 replace github.com/adcontextprotocol/adcp-go => ../../
 
-replace github.com/adcontextprotocol/adcp-go/adcp => ../../adcp
+replace github.com/adcontextprotocol/adcp-go/urlcanon => ../../urlcanon
